+++ processing/Neuro_DC/DoublePIDSimulator.pde
@@ -73,7 +73,7 @@ public class DoublePIDsimulator extends Simulator
       }
     }
   
-    if((Double)errorIntegral == Double.NaN) 
+    if(errorIntegral == Double.NaN) 
     {
       errorIntegral = Double.MAX_VALUE;
     }
